requirements.txt
ninja
yacs
cython
matplotlib
tqdm
Pillow